handle stacktrace frames without source code or line numbers
This uses essentially the same logic that StackTraceElement.toString() uses.

Note: the rollbar UI seems to ignore the value of `filename` when `lineno` is absent; it seems to always render `Unknown Source`. This is not terrible, but it means that the native method handling in this commit basically goes for naught.